+++ packages/mask/src/plugins/WalletService/messages.ts
@@ -1,6 +1,6 @@
 import type { AsyncVersionOf } from 'async-call-rpc'
 import { createPluginMessage, createPluginRPC } from '@masknet/plugin-infra'
-import { PLUGIN_ID } from './contants.js'
+import { PLUGIN_ID } from './constants.js'
 
 const WalletServiceMessages = createPluginMessage(PLUGIN_ID)
 export const WalletRPC: AsyncVersionOf<typeof import('./services/index.js')> = createPluginRPC(
