+++ packages/shared-base/src/Plugin/index.ts
@@ -47,5 +47,5 @@ export enum PluginID {
     ScamWarning = 'com.mask.scam-warning',
     SmartPay = 'com.mask.smart-pay',
     Claim = 'com.mask.claim',
-    WalletService = 'com.mask.walletservice',
+    WalletService = 'com.mask.WalletService',
 }
